[Java] fix: update test package following refactor
